requirements.txt
fastapi==0.114.0
uvicorn==0.30.6
google-generativeai==0.7.2
streamlit==1.38.0
